Reworked Upgrade card system, created easy way to create new cards
To create new cards, go to the Upgrade Cards folder and right click in the folder, select create and pick Upgrade card at the bottom. Fill in the fields for the cards you wanna create, to add the cards to the pool for available upgrades, select the EventSystem find the LevelUpManager component on it and drag the cards to the list of cards that already are there.